requirements.txt
python-dotenv==1.0.1
unstructured==0.11.8
faiss-cpu==1.8.0
flask==3.0.3
PyPDF2==3.0.1
langchain==0.1.20
langchain_community==0.0.38
langchain_google_genai==1.0.3
google-generativeai==0.5.2